modified makefile to work with floats printf
and other stuff in menu.c etc.

+++ freq_counter.h
@@ -2,4 +2,4 @@
 
 void freq_counter_init(void);
 
-uint32_t freq_counter_read(void);
+float freq_counter_read(void);
